Refactor: move network scan into separate package, minor code reformat after merge
