+++ drivers/network/info.go
@@ -1,4 +1,4 @@
-package shared
+package network
 
 import (
 	"net"
